% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_bed.R
\name{anno_bed}
\alias{anno_bed}
\title{Assign a position to genomic features by ChIPseeker}
\usage{
anno_bed(pos, reference = "hg38")
}
\arguments{
\item{pos}{three columns of the positions, chromosome,start,end}

\item{reference}{Choose hg19,hg38,mm10 ,defaults: hg38}
}
\value{
assigned results
}
\description{
filter SNP or INDELs in a vcf or maf file
}
\examples{
anno_bed
anno_bed(pos,reference='hg38' )
}
