% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_DM.R
\name{cal_DM}
\alias{cal_DM}
\title{Calculate DM(distance to median) values based on cpm(counts per million) expression matrix}
\usage{
cal_DM(exprSet, geneType = "ensembl", species = "human")
}
\arguments{
\item{exprSet}{cpm(counts per million) expression matrix}

\item{geneType}{Choose ensembl or symbol,defaults: ensembl}

\item{species}{Choose human or mouse,defaults: human}
}
\value{
log10cv2_adj DM(distance to median) values for each gene
}
\description{
expression matrix should be numeric values matrix in which  each row is a gene, and each column is a cell or sample.
The rownames for this expression matrix  should be ensembl IDs or gene symbols
One should set the species as human or mouse.
The gene length information calculated base onf GENCODE database.
}
\examples{
cal_DM
cal_DM(exprSet,geneType='ensembl',species='human')

}
