% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_ID_of_exprSet.R
\name{change_ID_of_expreSet}
\alias{change_ID_of_expreSet}
\title{Change the colnames of expression matrix between ensembl and symbol}
\usage{
change_ID_of_expreSet(exprSet, geneType = "ensembl", species = "human")
}
\arguments{
\item{exprSet}{The expression matrix}

\item{geneType}{Choose ensembl or symbol,defaults: ensembl}

\item{species}{Choose human or mouse,defaults: human}
}
\value{
exprSet The expression matrix which had been changed.
}
\description{
expression matrix should be numeric values matrix in which  each row is a gene, and each column is a cell or samples.
The rownames for this expression matrix  should be ensembl IDs or gene symbols
One should set the species as human or mouse.
The expression value will be sum if two or more genes to one gene
}
\examples{
change_ID_of_expreSet
change_ID_of_expreSet(exprSet,geneType='ensembl',species='human' )
}
