% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts2exprSet.R
\name{counts2exprSet}
\alias{counts2exprSet}
\title{transform and normlize the raw counts matrix}
\usage{
counts2exprSet(counts)
}
\arguments{
\item{counts}{The raw counts matrix from featureCounts or other tools}
}
\value{
exprSet, the normalized expression matrix, log2(cpm+1)
}
\description{
row counts matrix will be filtered and log2(cpm+1) tranformed.
}
\examples{
counts2exprSet
counts2exprSet(counts)
}
