% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprSet2cellcycle.R
\name{exprSet2cellcycle}
\alias{exprSet2cellcycle}
\title{Assign cell cycle status based on expression matrix}
\usage{
exprSet2cellcycle(exprSet, geneType = "ensembl", species = "human")
}
\arguments{
\item{exprSet}{The expression matrix(which shoud be normalized,like log2(cpm+1) or log2(tpm+1))}

\item{geneType}{Choose ensembl or symbol,defaults: ensembl}

\item{species}{Choose human or mouse,defaults: human}
}
\value{
assigned A list of (phases,scores,normalized.scores) return from cyclone(scran)
}
\description{
The rownames of expression matrix, should be ensembl IDs or gene symbols
One should set the species as human or mouse.
}
\examples{
exprSet2cellcycle
exprSet2cellcycle(exprSet,geneType='ensembl',species='human')
}
