% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprSet2CNV.R
\name{exprSet2CNV}
\alias{exprSet2CNV}
\title{Calculate CNV based on normalized expression matrix}
\usage{
exprSet2CNV(exprSet, geneType = "ensembl", species = "human")
}
\arguments{
\item{exprSet}{The expression matrix(which shoud be normalized,like log2(cpm+1) or log2(tpm+1))}

\item{geneType}{Choose ensembl or symbol,defaults: ensembl}

\item{species}{Choose human or mouse,defaults: human}
}
\value{
all_cnv CNV value matrix for each gene in each cell
}
\description{
expression matrix should be numeric values matrix in which  each row is a gene, and each column is a cell or sample.
The rownames for this expression matrix  should be ensembl IDs or gene symbols
One should set the species as human or mouse.
The genomic positions is human_geneInfo_genecode_v25 or mouse_geneInfo_genecode_vM12
}
\examples{
exprSet2CNV
exprSet2CNV(exprSet,geneType='ensembl',species='human')
}
