% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprSet2inferCNV.R
\name{exprSet2inferCNV}
\alias{exprSet2inferCNV}
\title{Create two input files for inferCNV based on expression matrix}
\usage{
exprSet2inferCNV(exprSet, geneType = "ensembl", species = "human",
  prefix = "example", dir = "./")
}
\arguments{
\item{exprSet}{The expression matrix(which shoud be normalized,like log2(cpm+1) or log2(tpm+1))}

\item{geneType}{Choose ensembl or symbol,defaults: ensembl}

\item{species}{Choose human or mouse,defaults: human}

\item{prefix}{The prefix for files,defaults:example}

\item{dir}{Choose where to put the files,defaults: ./}
}
\value{
Two files \code{example_inferCNV_pos.txt} and \code{example_inferCNV_exprSet.txt}
}
\description{
expression matrix should be numeric values matrix in which  each row is a gene, and each column is a cell or samples.
The rownames for this expression matrix  should be ensembl IDs or gene symbols
One should set the species as human or mouse.
Check the document for input format at : https://github.com/broadinstitute/inferCNV/wiki
The genomic positions for airway data is : human_geneInfo_genecode_v25
}
\examples{
exprSet2inferCNV
exprSet2inferCNV(exprSet,geneType='ensembl',species='human',prefix='airwary',dir=dir)
}
