% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genomic_positons.R
\name{get_genomic_positions}
\alias{get_genomic_positions}
\title{Get the genomic positions based on a list of gene}
\usage{
get_genomic_positions(geneList, geneType = "ensembl", species = "human")
}
\arguments{
\item{geneList}{Should be the rownames of expression matrix}

\item{geneType}{Choose ensembl or symbol,defaults: ensembl}

\item{species}{Choose human or mouse,defaults: human}
}
\value{
pos the genomic positions, columns should be : gene/chr/start/end
}
\description{
A list of gene, should be ensembl IDs or gene symbols
One should set the species as human or mouse.
The genomic positions is human_geneInfo_genecode_v25 or mouse_geneInfo_genecode_vM12
}
\examples{
get_genomic_positions
get_genomic_positions(rownames(exprSet),geneType='ensembl',species='human')
}
