% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_cnv.R
\name{ht_cnv}
\alias{ht_cnv}
\title{Draw heatmap for CNV matrix}
\usage{
ht_cnv(cnv, meta, prefix = "test", noise_filter = 0.2, upper = 2)
}
\arguments{
\item{cnv}{The CNV results from exprSet2CNV}

\item{meta}{Choose ensembl or symbol,defaults: ensembl}

\item{prefix}{The prefix of the filename of the PDF heatmap}

\item{noise_filter}{A value must be atleast this much more or less than the reference to be plotted [Default 0.2]}

\item{upper}{The maximum value for heatmap  [Default 2]}

\item{species}{Choose human or mouse,defaults: human}
}
\value{
heatmap
}
\description{
CNV matrix calculated base on expression matrix by exprSet2CNV
}
\examples{
ht_cnv
exprSet2CNV(exprSet,geneType='ensembl',species='human')
}
