% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam50subtyping.R
\name{pam50subtyping}
\alias{pam50subtyping}
\title{Pam50 subtyping based on normalized expression matrix}
\usage{
pam50subtyping(exprSet, geneType = "ensembl", species = "human")
}
\arguments{
\item{exprSet}{The expression matrix(which shoud be normalized,like log2(cpm+1) or log2(tpm+1))}

\item{geneType}{Choose ensembl or symbol,defaults: ensembl}

\item{species}{Choose human or mouse,defaults: human}
}
\value{
all_cnv molecular.subtyping results
}
\description{
expression matrix should be numeric values matrix in which  each row is a gene, and each column is a cell or sample.
The rownames for this expression matrix  should be ensembl IDs or gene symbols
One should set the species as human or mouse.
}
\examples{
pam50subtyping
pam50subtyping(exprSet,geneType='ensembl',species='human')
}
